backend/requirements.txt
# --- FastAPI Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.30.0
python-dotenv==1.0.1
google-genai==1.41.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1